Update DOCUMENTATION
DOCUMENTATION
The Ayesha Beauty Creation has introduced a software system that you can access even with your handy
mobile device through an ABC App. Basically the system is introduced to facilitate and automate the
business operations an workflow. The software system assists with transactions were a customer has to
select his/her products that he/she is interested in and the system automatically display the name and
the price of the products. Apart from that, other major role in financial return is that the system also
help with calculations of selling price, cost price and it output the financial return percentage.
The technology system also improve the efficiency in human resources where its main function is to
record the employee’s attendance. The system allows the employee to submit his or her attendance if
he or she is at work place and when the time is exactly 07:300 – 08: 00 each day from Monday to Friday,
else the system will proceed and the function to record absent and close itself.
The other important function of this system is that it help a market to create a database management
that keep its customer’s records and keep updating its customer whenever there is a sales, new stock
and new products available in the market. This makes it easier for a market to create a good relationship